+++ src/runner.ts
@@ -1,11 +1,9 @@
 import { URL } from 'url'
 import extOs from 'yyl-os'
-import { CommitConfig, Env, YylConfig } from 'yyl-config-types'
+import { CommitConfig, Env, YylConfig, Logger } from 'yyl-config-types'
 
-import { LANG } from './const'
 import { YServer, YServerSetting } from './server'
 import { YProxy, StaticFnOption, LocalRemote } from './proxy'
-import { Logger } from 'yyl-seed-base'
 
 export interface RunnerOption {
   yylConfig?: YylConfig